fix: request serializers 추가
